mpOpenApi-20: Initial multi-app support
Enable merging the OpenAPI documentation generated for multiple apps
deployed to the same server.

Users can configure which apps they want included in the merge.

In some circumstances, apps may clash and merging won't be possible. In
this case an error is emitted.

Unit tests for the merge logic are included.