mpOpenApi-20: Multi-app support
Enable merging the OpenAPI documentation generated for multiple apps
deployed to the same server.

Users can configure which apps they want included in the merge.

In some circumstances, apps may clash and merging won't be possible. In
this case an error is emitted.

Unit tests for the merge logic are included.

FAT tests are included which focus mainly on logic which can't be
covered by the unit tests.

Beta fenced so that non-beta runtimes are locked to the default config
which matches the current behaviour.