Issue 17072: Add support for JWT-formatted UserInfo responses
For #17072

Adds support for parsing UserInfo responses in JWT format, if found. The Content-Type header of the response is checked for `application/json` or `application/jwt` values and parsing is done accordingly.

Note that UserInfo responses in JWE format aren't required to contain a nested JWS, so the payload of a JWE could be raw JSON. The updated behavior also takes that into account.