mpOpenApi-20: Keep identical tags and components
When merging two documents which contain a component with the same name,
we usually rename the component in the second document. However, when
the two components are identical, we don't need to do this, we can just
have a single component in the merged document.

Tags are handled similarly. If the tag definition is identical, we don't
need to rename the tag. Ad-hoc tags used without definition at the top
level are never renamed.