mpOpenApi-20: Retain docs and info when identical
If the externalDocumentation and info sections of the original documents
are identical, retain them in the merged document. Otherwise they're
removed or replaced in the merged document.

Also refactor the processing steps slightly so that documents which
clash and are removed from the merge, can't cause things to be renamed
in the final merged documents.