mpOpenApi-20: Config source on app CL, not TCCL
Expose the DefaultsConfigSource on the app classloader, rather than the
TCCL.

This is important because we're exposing the source loader resources on
the app classloader. If for some reason we try to service load config
sources using the app classloader, it will fail because it sees the
service loader resource, but not the referenced class.