mpOpenApi-20: Fix test for use of config in filter
This test was changed when it was copied from the 1.x FAT to use the app
classloader, rather than the TCCL, which subverted the original purpose
of the test.

Restore the original test, but keep the additional test which uses the
app classloader directly as this also exposed problems with the way we
were providing our defaults config source.