mpOpenApi-20: Remove path context root heuristic
Remove the heuristic that if all paths start with the context root, then
the context root shouldn't be prepended. Enough tests hit this
unexpectedly that I think it does more harm than good, plus without
servers also present in the document, the paths in the document won't
line up with the actual endpoints.