Update jigawatts to return error codes from criu dump
Initially exit if criu dump returns an error.
It is unclear if we should exit only for criu dump errors
or all exceptions when doing a snapshot. For now
just simply exit when criu dump errors.

Much more work is needed to handle errors here.
For example, proper error messages and proper
FFDCs if they are appropriate.