Don't put exception stack in json error objects
When the JMX REST APIs encounter an error, the caught exception is serialized into
a JSON object which is returned from the API. The stack trace from the exception is
also added to the JSON object as a string field. This is not read by our REST client,
and was flagged as a security concern. Instead, put the exception message in a
string field to provide some human debuggability.