mpOpenApi-20: delegate methods where possible
Delegate methods within RenameReferenceVisitor where possible. I'm not
entirely sure about this change, but it does at least reduce the
duplication of code slightly and reinforces that the implementation of
both methods is the same. I'm not totally sure whether this is worth
adding the extra step of redirection or calling a method declared on the
interface in a different way to how it was designed. However, we did
spend way too long thinking about it.