Add org.eclipse.openj9.criu for boot delegation
This is not the ideal way to do this.  This package
is needed when running on Open J9 with CRIU support.

It would be better if the launcher would detect the
available packages from the JVM boot layer modules
similar to how the Equinox framework does when the
system packages are not set explicitly.

That is complicated to do given how we set certain
attributes on the system packages in the liberty
boot launcher.  For now just doing boot delegation
for this package.