Issue 16994: Update how dynamic reconfig of social login affects discovery data
Resolves #16994

This now sets the `discoveryDocumentHash` in `OidcLoginConfigImpl` back to null if the `discoveryEndpointUrl` is dynamically removed from the configuration or set to an empty string after discovery has already been performed. This should ensure that the various endpoint information in the discovery data gets re-populated if the `discoveryEndpointUrl` is dynamically re-added.