Avoid major version changes to utils package
The utils package is considered SPI. This limits
what we can change to the classes in the package
TimestampUtils contains a couple of protected fields.
Even though nobody would have a valid reason to
extend TimestampUtils we still need to make sure
the protected fields do not break compatibility.

Deprecating the protected fields and adding a new
private field for the start time.