Backup workarea for checkpoint, use backup on subsequent restores
When doing a checkpoint the workarea is backed up.  On restore a marker
file is created to indicate that on subsequent restores the workarea
should be restored before doing a criu restore.

This is done to preserve the workarea as it was when the checkpoint was
taken.  An environment variable is introduced (WLP_IN_CONTAINER) that
can be set to true.  When true then the workarea is not backed up on
checkpoint.  The assumption is that the container is always restarted
from the checkpoint image and therefore there is no need to backup the
workarea because the workarea saved in the container image will be
correct.