For checkpoint workarea backup, preserve the timestamps
Some caches in liberty are timestamp sensitive.  For example,
the library cache used for spring boot applications seems to
be senstive to any timestamp change of the libraries which
then cause the application to be restarted after a restore
from the backed up workarea.

Doing a cp --preserve=timestamps seems to work around this