Add deployment phase for checkpoint
The deployment checkpoint happens before executing
the first line of code from an application.
The deployment checkpoint is triggered by code
that is injected into application classes <clinit>
with a ClassFileTransformer.

Once the first <clinit> is done for any application
class then the checkpoint is performed.  When any
checkpoint is performed then the transformer is
unregistered to stop modifying application classes
on the restore side.