Add beta guard for the checkpoint action
This is done in LaunchArguments instead of in the server script
because a beta guard check cannot be efficiently done in the server
script.  As a result there is a hardcoded production of the message
about the checkpoint action being unknown.

This is because the server will not use the --help:actions:checkpoint
option when a GA driver is detected.