Fix handling of MP JWT default instance config
The `setGenericAndFilteredConfigTaiRequestInfoFromConfigServices()` method in the `TAIRequestHelper` class has some quirky logic for adding "generic" MP JWT configs to the `MicroProfileJwtTaiRequest` object. In short, if a server admin doesn't configure an `<mpJwt>` element, the default `<mpJwt>` element that's provided by the feature doesn't actually get added as a generic config for the request. That results in the MP JWT feature believing there are no configs present to handle a protected resource request, so Liberty moves on to the next authentication method. The default `<mpJwt>` element should be considered a generic config for this scenario.