Add whiteboard for ClassFileTransformer services
This allows a system transformer to be registered that applies
to all AppClassLoaders that get created for the server.

To be recognized as a system transformer the ClassFileTransformer
must be registered with the service property:

io.openliberty.classloading.system.transformer=true

System transformers differ from other transformers
added to the AppClassLoader class with
AppClassLoader.addTransformer(ClassFileTransformer). System
transformers can transform classes even if they were previously
stored in the shared classes cache.  But if a system transformer
does change a class bytes then the transformed class is not
stored in the shared classes cached, just like other transformers