mpOpenApi-2.0: return correct app path on clash
When clashes result in all but one OpenAPI model being excluded from
merging, the single model is returned with no processing done (exactly
the same as if there were no other apps deployed in the first place).

In this case, we need to ensure that the correct application path is
returned by the MergeProcessor as well.