Use Equinox internals to avoid CNFE on delegation
Equinox added an internal to avoid CNFE when
doing bundle delegation for class loading.
The idea is that generating exceptions is expensive
and should be avoided for highly used code flow.

This change needs to add a new internal interface
so we can know to call an internal method that returns
null for a class not found instead of throwing the
CNFE.  This allows us to delegate to the gateway
class loader more efficiently for API access by the
application class loader.