mpOpenApi-2.0: avoid processing apps on shutdown
In the default configuration, where only one app is processed to create
OpenAPI documentation, when that app is removed we then processed the
next app so that while any app is deployed we will have OpenAPI
documentation for one of them.

During shutdown all apps are removed in turn so we can end up creating
OpenAPI documentation for some apps while the server is shutting down,
which is a waste of time.