MP Config: improve performance of ServerXML vars
Improve performance when finding properties defined as variables in
server.xml by eliminating copying of maps when retrieving variables.

Note this changes the semantics of
ConfigVariables.getUserDefinedVariableDefaults. Previously the default
value of a variable would not be returned if it was overridden with an
actual value, whereas now it is.