Update OpenJ9 CRIU API with support for prepare and restore hooks
Now we can pass prepare and restore Runnables to the CRIUSupport so that
our prepare and restore hooks will be run by the JVM in single threaded
mode.

The JVM also changed how isCRIUSupportedEnabled works by having it
always return true or false and never throwing an exception.  Instead we
now have a method CRIUSupport.getErrorMessage() that can tell us why
CRIU is not supported.