Some hooks may required a JVM in multi-threaded mode
By default the hooks prepare and restore hooks are run
while the JVM is in single-threaded mode.  Some hooks
may require that they be running while the JVM is in
multi threaded mode.  This change adds a service
property for a hook to specify that they must be
run while the JVM is in multi-threaded mode.

The CheckpointOSGiConsole hook has been updated to be run
while the JVM is in multi-threaded mode and a new FAT
test has been added to ensure the osgiConsole-1.0 feature
is functional for checkpoint/restore