Resolver: test making install lists with tolerated
Test that install lists are correctly created when:
* A feature with a tolerated dependency is already installed
* The solution from the kernel resolver requires a new feature to be
  installed to fulfil that dependency

This situation can occur when there's a single public feature which
supports EE8 and EE9 by having a private feature for each and tolerating
either of those private features.

If, say, all the EE8 features are installed and the user requests to
install EE9, the private feature which supports EE9 needs to be
installed, even though the feature which has a dependency on it is
already installed.