Resolver: fix install lists for tolerated deps
Fix the creation of install lists when a feature is installed but its
dependencies aren't. This can happen when a feature tolerates two
versions of a dependency, only one of which is installed, and the
solution from the kernel resolver uses the other one.

This situation is common in Liberty where we have a single feature which
supports both EE8 and EE9 by having a private feature for each and
tolerating both of them. The eeCompatible singleton features will ensure
that the correct one is used, depending on which other features are
enabled on the server.

This fix includes the following changes:
* When building an install list, recurse through the dependencies of any
  installed features rather than stopping. The installed features are
  removed later when features are mapped to EsaResources.
* Refactor the code for finding the features used to satisfy the
  ProvisionCapability header of an auto-feature so that it can be used
  for features which are installed as well as features in the
  repository.
* When building an install list for samples, don't try to build an
  install list for a feature which wasn't found.