Resolver: fix install list for unresolved features
If a feature requested by the user isn't resolved by the kernel
resolver, ensure we don't try to make an install list for it and that
a suitable error is returned.

This can happen when the user directly requests two conflicting
features because neither is returned by the kernel resolver.