Resolver: remove redundant checks
The sets featuresMissing and requirementsFoundForOtherProducts are
disjoint. If we're checking that featuresMissing contains a feature
name, we don't need to also check that requirementsFoundForOtherProducts
doesn't contain it.

Previously we just had a check to ensure it wasn't in
requirementsFoundForOtherProducts. This wasn't correct either as it's
possible for the feature to be in neither set if it conflicted with
another feature.