Issue 20165: Update the json container features to include the BELLS feature.
Update the json container features to include the BELLS feature and remove Liberty's default implementations of JSON-P and JSON-B. This allows for a user to include their own third party library for these JSON technologies. Additionally this PR updates the JSON container FAT tests to run against the 2.0 (Jakarta EE 9) to verify this fixed functionality.