CDI 4.0: Allow BCE BDA to see app beans
The Lite Extension Translator implements Build Compatible Extensions at
runtimes. One of the requirements for BCE is that a SyntheticBeanCreator
is passed an Instance<Object> which it can use to lookup other beans
from the application. Therefore, the BDA which holds the Lite
Extension Translator needs to be able to see application beans.