CDI 4.0: Don't discover beans when there's a BCE
In CDI, archives which contain an extension are not bean archives unless
they also contain a beans.xml. With CDI 4.0, the concept of "extension"
includes build compatible extensions.

We need to ensure that we don't discover beans within archives with a
BCE, unless there's also a beans.xml. We do this by making sure that
WebSphereBeanDeploymentArchive.isExtension() returns true if there's
either a BCE or a portable extension.

To do that, we make BDAs and Archives aware of BCE class names, but we
delegate the finding of those class names to a BceFinder service, which
we only implement in CDI 4.0.

As part of the refactoring, some of the logic that was in
LiteExtensionProvider is now in CDI40BceFinder.