Use findClassNoException to get bootdelegation
Fixes issue #20693

With the calls to BundleLoader.findClassNoParentNoException
we no longer get "free" bootdelegation when delegating
to the bundle loader.  This fix instead delegates
a new method findClassNoException which will delegate
for bootdelegation.

Note that this is using an overlay to add the method
to BundleLoader.  This method is not in Equinox yet.