Improve JWT cache performance
Cache behavior is updated to not re-validate JWTs that have already been validated. The behavior before was looking up the `JwtContext` from the cache but was still re-validating JWTs and therefore wouldn't be much of a performance improvement. This also moves where we cache the JWT context to after we perform the actual validation of the token.