Remove vendor check
The check is largely not necessary here because tests that
follow this codepath would likely have the necessary
skip annotations to avoid running the test on a JVM
that cannot support checkpoint.

Additional changes may be considered to add these
simple vendor checks to the skip logic to make them
faster by avoiding the more involved probe checks
there.