Add option to avoid auto-recovery when criu restore fails
Add option CRIU_RESTORE_DISABLE_RECOVERY that when not set to
a value then the server script will attempt to recover from
a failed restore with a normal server launch.

Note that this currently only handles when criu itself fails
to restore the process.  Additional work is needed to handle
when the JVM or Liberty restore hooks fail.