Implemented FeatureIntrospector to output server feature definitions
In the event of a dump, the FeatureIntrospector will iterate over all server features and output the feature name, as well as its visibility, version, simple name, and symbolic name. If the feature is a provisioning feature, then the bundle repository type, auto feature, and singleton status will be output. Whether or not the feature is superseded, and what it is superseded by will be output, as well as its constituents and thier associated attributes.