Wait for kafka to finish creating topic
Calling kafkaClient.partitionsFor(topic) should create the topic if it
doesn't exist, but in some cases I'm seeing failures because it doesn't
and instead returns null or an empty list of partitions.

This may be because the topic is being created but the creation hasn't
finished yet, it seems to happen intermittently.

Since this failure happens in the test setup, I'm happy to just add a
retry with a timeout here.