Add new test for non-persistent timers and checkpoint
This test adds a timer to the test EJB application.
The timer goes off every second.  It keeps track that it is
not called too fast indicating some catchup calls on
restore.  The expected result is that no catchu calls
are done on restore no matter how long the time is
between the checkpoint and the restore process.
The test waits 5 seconds before restoring the process
which would cause 5 timers to be missed.