Issue 20356: Add jti check to back-channel logout token validation
Verifies that logout tokens with identical `jti` claims cannot be reused within a relatively short time frame. The code is currently written for that time frame to be 10 minutes. Logout tokens being processed are tied to the OIDC client config doing the processing, which means two logout tokens with the same `jti` should still be usable by two separate OIDC clients.

For #20356