Refactor OidcSessionInfo related classes
Updates the `OidcSessionCache` interface (and `InMemoryOidcSessionCache` implementing class) to more directly use `OidcSessionInfo` data when caching. Instead of caching just the session ID, we'll cache the whole `OidcSessionInfo` object. That makes it a bit more straightforward to work with in memory rather than having to re-parse the data from the session ID.

The `OidcSessionHelper` class is also deleted in favor of moving its logic into `OidcSessionInfo`.