Modified checkpoint command to avoid writing to log file if server does not exist
In the event of a checkpoint, the server script will check if the server.xml file exists in the SERVER_CONFIG_DIR. If the specified server is non-existant, then the console.log file will not be created. However, the server working directory will always be created whether or not the server is existant, this behavior is consistent with the "run" and "debug" actions.