Issue 21376: Remove access and refresh tokens from the OP cache
Removes any ID, access, and refresh tokens that are associated with the user and RP that is being logged out as part of OIDC back-channel logout. All of the user's cached tokens are examined to find ID and access tokens associated with the user and client, and then any refresh token that is associated with each of those tokens is also removed from the cache.

Resolves #21376