Add methods to CheckpointPhase to add hooks
To make it easier for static code that needs
to participate in checkpoint a new static getPhase
method is added to allow static code to get the current
phase that is being used for checkpoint.  If no checkpoint
is being done then null is returned.  In addition two
new methods were added to the CheckpointPhase instance
to allow hooks to be added to the current phase

This way static code can participate with no need
to have a BundleContext.  Something like this would
likely be needed if we extend the checkpoint as API for
applications to participate.