Add io.openliberty.checkpoint.spi package to allowed list
This test is flaky and likely not testing what it is
intended to test.  It calls ClassLoader.getPackages to
determine what packages are available from the JVM
app class loader.  This is not a reliable way to
determine that because it depends on at least
one class from each and every package available
to have been loaded before this call in order
to have the actual Package objects created.

This forces us to add new packages sometimes which
were already available, but never loaded yet.