Allow EnvConfigSource to work with checkpoint
Call super to get the list of defaults then replace
the instance of EnvConfigMap with another one that
doesn't make a copy of the env Map.

This allows us to avoid making a copy of the System.getenv Map
so that environment changes on restore can be accurately reflected
in mpConfig on restore.