Added a checkpoint with method timeout annotation test to checkpoint_fat FATSuite
Created a testcase to check the method timeout behavior when a timeout is invoked during a checkpoint. When the checkpoint command is run, the server will spawn a thread which starts a timeout. When the server is restored, the timeout method should resume ignoring time passed in which the server was inactive. Added an additional test to check the behavior when a timeout does not occur at checkpoint. Also added sleep method in between checkpoint and restore to ensure a period greater than the expected timeout has passed.