Update CPUInfo to handle Semeru version 8 returning IBM as the vendor
- Since JavaInfo.vendor() returns IBM for Semeru 8 we were thinking that
the CPU time need to be adjusted because the Java version was Java 8 SR
0 because the version is 1.8.0_xxx-bxx for Semeru instead of 1.8.0 SRx
FPy.  As such needed to adjust how we detect that the JVM is IBM Java 8
- Updated to add a toString to CPUCount so you can see the CPU count in
the trace.
- Added a test so that we can easily test the bad behavior with
different versions of Java.  It is only a unit test though so it is part
of the build or run in your IDE.