Issue 21447: Extract user name from session security name in BCL
Before, the OIDC `BackchannelLogoutService` class expected the user name it received to be the correct string as-is. However the user name that is returned from the `IHttpSessionContext` can take the format of an access ID. We therefore need to extrac the user name from that access ID if the input string is in that format.

Resolves #21447