Check that system WABs are provisioned before checkpoint
System WABs do hold up the server started event, but they currently do
not block a checkpoint if the background thread is not done processing
the system WABs from the web container.  It would be best to make sure
that async work is done before proceeding with a checkpoint to avoid
potential deadlock while in single threaded mode and also to reduce the
work necessary on restore.