Return boolean from add hook methods
Timing may prevent a hook from being added when a checkpoint is being
done.  This can happen if a background thread gets the CheckpointPhase
and checks the restore status.  If the restore status is false then they
can try adding the hook.  But before adding the hook the process may
have moved past checkpoint.  To eliminate the timing issue the add hook
methods are updated to return boolean if the hook was added
successfully.  For cases where the hook did not get added successfully
then the caller can decide to do something else.