Only pay attention to the timer delta after the first delta calculation
Seems the very first call to the timer may be not aligned directly
with a second making it appear that the second call to the timer
happens too early depending on the offset from the second
boundary.  This change only looks at the delta time between
calls to the timer after the first delta time is calculated.