Updated the RESTclient application to use the baseURI defined in the server.xml file
The server will first start normally and check that the default baseURI defined in the testClient variable is used. Then, the server will checkpoint and change the testClient variable in the server.xml file to the alternate endpoint. After restoring and verifying that the alternate endpoint was accessed, the default endpoint will be restored via the setHost endpoint.