Issue 21766: Do not make BCL calls for public clients with HTTP BCL URIs
Updates the OIDC back-channel logout code on the server side to validate the client's back-channel logout URI before adding it as a client to send BCL requests to. The code re-uses the same logic used to validate the BCL URI of an OAuth client that is being created or updated.

Resolves #21766