Fix app validation logic for started servers
The previous logic was broken:

When the first app (e.g. named appOne) is deployed, the
startApplicationMessages counter is incremented and it waits for one
message matching CWWKZ0001I:.*appOne

When the second app (e.g. named appTwo) is deployed, the
startApplicationMessages counter is incremented and it waits for *two*
messages matching CWWKZ0001I:.*appTwo

It'll never get two messages so it will always wait for the full
timeout, log an error and then continue.

Instead, this change means we will look for messages indicating that the
application has either started or stopped and assumes that the most
recent message represents the application's current state.

If there are no started or stopped messages, the application is assumed
to be stopped.