Fix app deploy validation in wsoc FAT
Although this test deploys an app to dropins, the server configuration
disables monitoring of dropins so the app is not expected to start and
we need to use the DISABLE_VALIDATION flag.

Later on the server.xml is updated to list this archive as an app to be
run and the log is checked at that point to make sure it starts.