Update app start validation in JWT FATs
Previously the JWT FATs were removing an app for validation and then
calling LibertyServer.reconfigureServerUsingExpandedConfiguration to
update the configuration and validate the expected apps start. This
doesn't work because when you remove an app for validation from a
running server, the method immediately checks that the app is not
running.

However, doing these steps in the reverse order doesn't work either
because LibertyServer.reconfigureServerUsingExpandedConfiguration checks
that all apps added for validation start after the new server config is
applied, so that would fail if the list of apps added for validation
isn't changed first.

As a solution, I've avoided using
LibertyServer.reconfigureServerUsingExpandedConfiguration at all
whenever we're changing the configuration in a way that would install or
uninstall apps. This also allows us to avoid changing the list of apps
added for validation so that the cleanup logic which resets the
configuration after every test still works.