Fix app validation in servlet 4.0 FAT
A test was deploying an app and then updating the config to start the
app, therefore we need to use DISABLE_VALIDATION when the app is
deployed because it's not expected to immediately start.

We could add it for validation after the config is updated, but then
we'd have to ensure we remove it again which would mean reverting the
config as part of a cleanup step for this test. It's easier to never add
it and allow the general cleanup logic to revert the config for us.