MPFT 1.x: Fix timeout cancellation race condition
Don't attempt to cancel the timeout task if we've already timed out.

There's a lock shared between stop() and timeout(), but if the timeout
has fired, the timeout future is not complete at the point where
timeout() releases the lock. This gives stop() a chance to run and
attempt to cancel the future. Doing this can result in an
InterruptedException which causes and FFDC to be emitted by the policy
executor.

By checking the value of timedout, we can avoid trying to cancel the
timeout task if the timeout has already fired and interrupted our
thread.