Configure dataSource after checkpoint
This test requires some placeholder variables be used to
set some placeholder values that are used during checkpoint.
Then the server.env is updated with the information to
connect to the actual database after checkpoint and before
restore.  The test validates that the updated datasource
configuration can be used to successfully connect to the
remote datasource.