Create internal xmlbinding 4.0 feature
- Create internal version of xmlBinding-4.0 that can be used by other
features to just get the runtime without exposing the API or the
XMLBinding tools
- Update features to use the XMLBinding 4.0 internal feature when it is
only used internally by the feature to read XML deployment descriptors.
This will allow for the web profile features to not expose the XML
Binding APIs.
- Update xmlBinding-4.0 feature so that the tools will work when doing a
minify since the tools depend on XMLBinding 3.0 versions of the jars.