Added "overwrite" deployment option to ShrinkHelper methods in test applications
The DefaultApp and ExportToServer methods of the ShrinkHelper class include the overwrite option in the BasicServletTest, CheckpointPhaseTest, LogsVerificationTest, MPFaultToleranceTimeouTest, OSGiConsoleTest, and RESTclientTest classes. This ensures that the exported archive will be overwritten if it exists. This is necessary for running repeated tests with test classes that use the same application.