Ensure logs are restored on error
If an error occurs during a checkpoint by criu
then the restore hook methods are not always run.

We need to make sure to restore at least the logs
to allow messages and trace logs to resume.
This is not ideal solution.  We should look
to adding a recover type call to the hooks if an
error occurs.