Rest client FAT review comments for #22141
Although ProduceConsumeTest specified that it ran on MP20 which is based
on EE8, it didn't actually have any EE features in the server.xml so it
would actually have run on EE7. Now that it has servlet in the
server.xml, update the test so that it runs with mpRestClient-1.1 on
both EE7 and EE8.

Fix the header.propagation server.xml to use servlet-3.1 to match the
other EE7 features listed. This didn't cause a failure because the
features in the server.xml will be overwritten by the repeat action in
the test, but it's confusing to have a mismatch of features.