Fix issues with finding the test method name.
The componenttest.topology.utils.FATServletClient.getTestMethodSimpleName(TestName)
sometimes returns the class-name.method-name.

This fixes that plus fixes an issue with the LocalEJBTest
that was having issues finding the proper TestMethod.

io.openliberty.checkpoint.fat.FATSuite.getTestMethod(Class<T>, TestName)
was updated to fail the test if a proper TestMethod enum
could not be found.