Allow thread dump before checkpoint
Added a system property "io.openliberty.checkpoint.dump.threads".
When set to any value it will cause a java thread dump
to be done while the JVM is in single-threaded mode
before the checkpoint is done.

The option -Xsyslog:none is needed to prevent the use
of syslog while doing a thread dump; otherwise the syslog
resource is left open, causing the checkpoint to fail.