Improve performance for generating the .env.properties
The script code for creating a valid properties file has
to account for multi-line env values.  The current approach
is to only include the first line of such values.

The previous code was taking longer than expected. This
change simplifies the code by unsetting INVOKED so we
don't have to filter it and then using a regex grep
to only include lines with proper key=value format.