Improve MPConfigTest to shorten the test time
The test was waiting too long for expected messages.  By the
time the test is checking the logs the warning for changed
config value will either have happened or not.  No need to
wait for the default timeout.

Also improved the test to ensure the server is always shutdown,
even on failures.