Issue 21157: Perform discovery for every OIDC request, not just for missing authz endpoint
Updates the Jakarta Security code to perform OIDC Discovery right at the start of each flow. Before, we were only performing discovery when the authorization endpoint URL wasn't configured in the provider metadata in the servlet annotation. This will ensure we always have the OP's metadata and that values in the provider metadata can be used to override values from the discovery data.

For #21157